trapAuthをauthServiceの機能に移動
chatの中で使いたいので、authServiceを受け取るよう変更
