+++ backend/src/chat/chat.gateway.ts
@@ -1,4 +1,3 @@
-import { JwtService } from '@nestjs/jwt';
 import {
   MessageBody,
   ConnectedSocket,
@@ -8,7 +7,6 @@ import {
   OnGatewayConnection,
 } from '@nestjs/websockets';
 import { Socket, Server } from 'socket.io';
-import { jwtConstants } from 'src/auth/auth.constants';
 import { ChatroomsService } from 'src/chatrooms/chatrooms.service';
 import { OperationGetRoomMembersDto } from 'src/chatrooms/dto/operation-get-room-members';
 import { OperationGetRoomMessageDto } from 'src/chatrooms/dto/operation-get-room-message';
@@ -16,7 +14,6 @@ import { OperationJoinDto } from 'src/chatrooms/dto/operation-join.dto';
 import { OperationLeaveDto } from 'src/chatrooms/dto/operation-leave.dto';
 import { OperationOpenDto } from 'src/chatrooms/dto/operation-open.dto';
 import { OperationSayDto } from 'src/chatrooms/dto/operation-say.dto';
-import { UsersService } from 'src/users/users.service';
 import { ChatService } from './chat.service';
 import * as Utils from 'src/utils';
 import { OperationKickDto } from 'src/chatrooms/dto/operation-kick.dto';
@@ -30,6 +27,7 @@ import {
   usersJoin,
   sendResultRoom,
 } from 'src/utils/socket/SocketRoom';
+import { AuthService } from 'src/auth/auth.service';
 
 const secondInMilliseconds = 1000;
 const minuteInSeconds = 60;
@@ -50,8 +48,7 @@ export class ChatGateway implements OnGatewayConnection {
   constructor(
     private readonly chatService: ChatService,
     private readonly chatRoomService: ChatroomsService,
-    private readonly jwtService: JwtService,
-    private readonly usersService: UsersService
+    private readonly authService: AuthService
   ) {}
 
   /**
@@ -59,7 +56,7 @@ export class ChatGateway implements OnGatewayConnection {
    * @param client
    */
   async handleConnection(@ConnectedSocket() client: Socket) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -127,7 +124,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationOpenDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -172,7 +169,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationSayDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -223,7 +220,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationJoinDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -315,7 +312,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationLeaveDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -359,7 +356,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationNomminateDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -417,7 +414,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationKickDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -474,7 +471,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationBanDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -527,7 +524,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationMuteDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -584,7 +581,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationGetRoomMessageDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -616,7 +613,7 @@ export class ChatGateway implements OnGatewayConnection {
     @MessageBody() data: OperationGetRoomMembersDto,
     @ConnectedSocket() client: Socket
   ) {
-    const user = await this.trapAuth(client);
+    const user = await this.authService.trapAuth(client);
     if (!user) {
       return;
     }
@@ -632,41 +629,6 @@ export class ChatGateway implements OnGatewayConnection {
         client,
       }
     );
-  }
-
-  //TODO 外部に出したい
-  private async trapAuth(client: Socket) {
-    if (client.handshake.auth) {
-      const { token, sub } = client.handshake.auth;
-      // token による認証
-      if (token) {
-        const verified = this.jwtService.verify(token, {
-          secret: jwtConstants.secret,
-        });
-        // console.log(verified);
-        const decoded = this.jwtService.decode(token);
-        if (decoded && typeof decoded === 'object') {
-          const sub = decoded['sub'];
-          if (sub) {
-            const userId = parseInt(sub);
-            const user = await this.usersService.findOne(userId);
-            if (user) {
-              return user;
-            }
-          }
-        }
-      }
-      // subによる認証スキップ
-      // TODO: 提出時には絶対に除去すること!!!!
-      if (sub) {
-        const userId = parseInt(sub);
-        const user = await this.usersService.findOne(userId);
-        if (user) {
-          return user;
-        }
-      }
-    }
-    return null;
   }
 
   private socketsInUserChannel(userId: number) {
